requirements.txt
cryptography
pyTelegramBotAPI
MySQL-python
